follow the newest dir layers
