+++ src/unit_tests/kubernetes_api_objects/mod.rs
@@ -1,3 +1,6 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
+pub mod config_map;
+pub mod container;
 pub mod object_meta;
+pub mod resource_requirements;
