+++ src/unit_tests/mod.rs
@@ -1,8 +1,4 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
-pub mod config_map;
-pub mod container;
 pub mod kubernetes_api_objects;
-pub mod object_meta;
-pub mod resource_requirements;
 pub mod vstd_ext;
